+++ purenessscopeserver/purenessscopeserver/PurenessScopeServer/Message/WorkThreadAI.cpp
@@ -211,7 +211,8 @@ uint16 CWorkThreadAI::GetReturnDataLength()
 
 void CWorkThreadAI::Close()
 {
-	uint32 u4Size = m_vecCommandTime.size();
+    uint32 u4Size = (uint32)m_vecCommandTime.size();
+
     for(uint32 i = 0; i < u4Size; i++)
     {
         SAFE_DELETE(m_vecCommandTime[i]);
@@ -242,7 +243,8 @@ void CWorkThreadAI::GetAllTimeout(uint32 u4ThreadID, vecCommandTimeout& objTimeo
 {
     //objTimeout.clear();
 
-	uint32 u4Size = m_vecCommandTime.size();
+    uint32 u4Size = (uint32)m_vecCommandTime.size();
+
     for(uint32 i = 0; i < u4Size; i++)
     {
         _CommandTime* pCommandTime = m_vecCommandTime[i];
@@ -270,7 +272,8 @@ void CWorkThreadAI::GetAllForbiden(uint32 u4ThreadID, vecCommandTimeout& objForb
 {
     //objForbiden.clear();
 
-	uint32 u4Size = m_vecCommandTimeout.size();
+    uint32 u4Size = (uint32)m_vecCommandTimeout.size();
+
     for(uint32 i = 0; i < u4Size; i++)
     {
         _CommandTimeout objData;
